🧵 feat: split Solidity resolution benchmarks into separate files
Split the monolithic resolutions.bench.ts into two specialized benchmark files:
- resolutions-single-file.bench.ts: Measures performance of single file imports
- resolutions-module-graph.bench.ts: Measures full module graph resolution

This separation provides clearer benchmarking focus and better organized test files.
