+++ test/bench/src/resolutions/resolutions-module-graph.bench.ts
@@ -1,15 +1,9 @@
-// DEPRECATED: This file has been split into two separate benchmark files:
-// - resolutions-single-file.bench.ts - For single file import resolution benchmarks
-// - resolutions-module-graph.bench.ts - For full module graph resolution benchmarks
-//
-// Please use those files instead of this one.
-
 import { bench, describe, expect } from "vitest";
 import path from "path";
 import fs from "fs";
 import { promises as fsPromises } from "fs";
 import { runPromise } from "effect/Effect";
-import { moduleFactory, resolveImports } from "@tevm/resolutions";
+import { moduleFactory } from "@tevm/resolutions";
 import { resolutionsRsFfi, isUsingNativeModule } from "./resolutions-rs-ffi.js";
 
 // Paths to our test fixtures
@@ -18,20 +12,6 @@ const MAIN_CONTRACT = path.join(FIXTURE_DIR, "fixture.sol");
 
 // Cache the file content to avoid disk I/O during benchmarks
 let mainContractCode = "";
-
-/**
- * Benchmarks the JavaScript implementation of resolveImports (for validation)
- */
-async function jsResolution(filePath: string, code: string) {
-  return runPromise(resolveImports(filePath, code, {}, [], false));
-}
-
-/**
- * Benchmarks the Rust implementation of resolveImports via FFI (for validation)
- */
-async function rustResolution(filePath: string, code: string) {
-  return resolutionsRsFfi.resolveImports(filePath, code, {}, []);
-}
 
 /**
  * Benchmarks the JavaScript implementation of moduleFactory
@@ -54,48 +34,6 @@ async function rustModuleFactory(filePath: string, code: string) {
 }
 
 /**
- * Loads all files in a directory recursively
- */
-async function loadAllFiles(
-  dir: string,
-  fileList: string[] = [],
-): Promise<string[]> {
-  const items = await fsPromises.readdir(dir, { withFileTypes: true });
-
-  for (const item of items) {
-    const fullPath = path.join(dir, item.name);
-
-    if (item.isDirectory()) {
-      fileList = await loadAllFiles(fullPath, fileList);
-    } else if (item.isFile() && item.name.endsWith(".sol")) {
-      fileList.push(fullPath);
-    }
-  }
-
-  return fileList;
-}
-
-/**
- * Validates that both implementations resolve imports correctly
- */
-async function validateImportResolution(filePath: string) {
-  const code = await fsPromises.readFile(filePath, "utf-8");
-
-  // Run both implementations
-  const jsResults = await jsResolution(filePath, code);
-  const rustResults = await rustResolution(filePath, code);
-
-  // Check that both implementations return the same number of imports
-  expect(jsResults.length).toBe(rustResults.length);
-
-  // Log the results for verification
-  console.log("JS Results:", jsResults.map((i) => i.original).sort());
-  console.log("Rust Results:", rustResults.map((i) => i.original).sort());
-  
-  return { jsResults, rustResults };
-}
-
-/**
  * Validates that both implementations build complete module graphs correctly
  */
 async function validateModuleFactory(filePath: string) {
@@ -109,21 +47,6 @@ async function validateModuleFactory(filePath: string) {
   console.log(`JS Module Graph Size: ${jsModuleMap.size} modules`);
   console.log(`Rust Module Graph Size: ${rustModuleMap.size} modules`);
   
-  return { jsModuleMap, rustModuleMap };
-}
-
-// Pre-load our main contract code and validate implementations
-async function setup() {
-  mainContractCode = await fsPromises.readFile(MAIN_CONTRACT, "utf-8");
-
-  // Verify import resolution for validation
-  console.log("Validating import resolution...");
-  await validateImportResolution(MAIN_CONTRACT);
-  
-  // Verify module factory
-  console.log("\nValidating module factory...");
-  const { jsModuleMap, rustModuleMap } = await validateModuleFactory(MAIN_CONTRACT);
-  
   // Print module graph details
   console.log("\nJS Module Graph:");
   for (const [path, module] of jsModuleMap.entries()) {
@@ -134,6 +57,17 @@ async function setup() {
   for (const [path, module] of rustModuleMap.entries()) {
     console.log(`- ${path} (${module.importedIds.length} imports)`);
   }
+  
+  return { jsModuleMap, rustModuleMap };
+}
+
+// Pre-load our main contract code and validate implementations
+async function setup() {
+  mainContractCode = await fsPromises.readFile(MAIN_CONTRACT, "utf-8");
+
+  // Verify module factory
+  console.log("Validating full module graph resolution...");
+  await validateModuleFactory(MAIN_CONTRACT);
 
   // Display which implementation we're actually using
   console.log(`\n=============================================`);
@@ -151,7 +85,7 @@ async function setup() {
   console.log(`=============================================\n`);
 }
 
-describe("Solidity Import Resolution Benchmark", async () => {
+describe("Solidity Full Module Graph Resolution Benchmark", async () => {
   // Initialize before running benchmarks
   await setup();
 
@@ -165,14 +99,5 @@ describe("Solidity Import Resolution Benchmark", async () => {
   const rustImpl = isUsingNativeModule ? "Native Rust" : "Simulated Rust";
   bench(`${rustImpl} - Full Module Graph Resolution`, async () => {
     await rustModuleFactory(MAIN_CONTRACT, mainContractCode);
-  });
-  
-  // Single file import resolution benchmarks (for comparison)
-  bench("JavaScript - Single File Import Resolution", async () => {
-    await jsResolution(MAIN_CONTRACT, mainContractCode);
-  });
-  
-  bench(`${rustImpl} - Single File Import Resolution`, async () => {
-    await rustResolution(MAIN_CONTRACT, mainContractCode);
   });
 });
